+++ src/solvers/adaptive_regularization_with_cubics.jl
@@ -166,15 +166,15 @@ function show(io::IO, arcs::AdaptiveRegularizationState)
     return print(io, s)
 end
 
-_doc_ARC_mdoel = raw"""
+_doc_ARC_model = """
 ```math
-m_k(X) = f(p_k) + ⟨X, \operatorname{grad} f(p^{(k)})⟩ + \frac{1}{2}⟨X, $(_tex(:Hess)) f(p^{(k)})[X]⟩ + \frac{σ_k}{3}\lVert X \rVert^3
+m_k(X) = f(p_k) + $(_tex(:inner, "X", "$(_tex(:grad)) f(p^{(k)})")) + $(_tex(:frac, "1", "2)) $(_tex(:inner, "X", "$(_tex(:Hess)) f(p^{(k)})[X]")) + $(_tex(:frac, "σ_k", "3"))$(_tex(:norm, "X"))^3"))
 ```
 """
 
-_doc_ARC_improvement = raw"""
+_doc_ARC_improvement = """
 ```math
-  ρ_k = \frac{f(p_k) - f(\operatorname{retr}_{p_k}(X_k))}{m_k(0) - m_k(X_k) + \frac{σ_k}{3}\lVert X_k\rVert^3}.
+  ρ_k = $(_tex(:frac, "f(p_k) - f($(_tex(:retr))_{p_k}(X_k))", "m_k(0) - m_k(X_k) + $(_tex(:frac, "σ_k", "3"))$(_tex(:norm, "X"))^3"))
 ```
 """
 _doc_ARC_regularization_update = raw"""
@@ -198,7 +198,7 @@ _doc_ARC = """
 
 Solve an optimization problem on the manifold `M` by iteratively minimizing
 
-$_doc_ARC_mdoel
+$_doc_ARC_model
 
 on the tangent space at the current iterate ``p_k``, where ``X ∈ $(_math(:TpM; p = "p_k"))`` and
 ``σ_k > 0`` is a regularization parameter.