+++ Changelog.md
@@ -16,6 +16,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   this is the dynamic variant to be used _after_ a solver has stopped.
 * a `has_converged(::AbstractManoptSolverState)` function to check whether the solver has converged.
 
+### Changed
+
+* formerly a stopping criterion could be activated at certain iterations with `sc > 5`, `sc >= 5`, `sc == 5`, `sc <= 5`, and `sc < 5`.
+  This caused too many issues with invalidations, so it has been reduced and moved to `sc ⩼ 5`, `sc ≟ 5`, `sc ⩻ 5` for the cases 1, 3, and 5, respectively,
+  cf. (#509).
+
 ## [0.5.22] September 09, 2025
 
 ### Added